rename golden to testdiff
